move code from signature/jws to signature
